change filename to snake_case
