+++ backend/api/models/assignment_rule_criteria.py
@@ -2,7 +2,7 @@ from django.conf import settings
 from django.db import models
 from django.utils import timezone
 from simple_history.models import HistoricalRecords
-from .assignmentRule import AssignmentRule
+from .assignment_rule import AssignmentRule
 from api.utils.model_manager import ActiveManager
 
 class AssignmentRuleCriteria(models.Model):
