+++ backend/api/models/assignment_rule_resource_group.py
@@ -1,26 +1,21 @@
 from django.conf import settings
 from django.db import models
 from django.utils import timezone
-from .operationalExceptionType import OperationalExceptionType
 from simple_history.models import HistoricalRecords
+from .assignment_rule import AssignmentRule
+from .resource_group import ResourceGroups
 from api.utils.model_manager import ActiveManager
-from .weeklyShiftTemplate import WeeklyShiftTemplate
 
-
-class OperationalException(models.Model):
+class AssignmentRuleResourceGroup(models.Model):
     id = models.AutoField(primary_key=True)
-    external_id = models.CharField(max_length=250, blank=True, null=True)
-    operational_exception_type = models.ForeignKey(OperationalExceptionType, on_delete=models.DO_NOTHING, blank=True, null=True)
-    start_datetime = models.DateTimeField(blank=True, null=True)
-    end_datetime = models.DateTimeField(blank=True, null=True)
-    notes = models.TextField(blank=True, null=True)
-    weekly_shift_template = models.ForeignKey(WeeklyShiftTemplate, on_delete=models.DO_NOTHING, blank=True, null=True)
+    assignment = models.ForeignKey(AssignmentRule, on_delete=models.CASCADE, related_name="assignment_resource_ids", blank=True, null=True)
+    resource = models.ForeignKey(ResourceGroups, on_delete=models.CASCADE, related_name="resource_assignment_ids", blank=True, null=True)
     
     # Metadata
     created_at = models.DateTimeField(default=timezone.now)
     created_by = models.ForeignKey(
         settings.AUTH_USER_MODEL,
-        related_name="created_operational_exception",
+        related_name="created_assignment_rule_resource_group",
         on_delete=models.SET_NULL,
         null=True,
         blank=True,
@@ -28,7 +23,7 @@ class OperationalException(models.Model):
     updated_at = models.DateTimeField(auto_now=True)
     updated_by = models.ForeignKey(
         settings.AUTH_USER_MODEL,
-        related_name="updated_operational_exception",
+        related_name="updated_assignment_rule_resource_group",
         on_delete=models.SET_NULL,
         null=True,
         blank=True,
@@ -36,15 +31,16 @@ class OperationalException(models.Model):
     deleted_at = models.DateTimeField(blank=True, null=True)
     is_active = models.BooleanField(default=True)
     is_deleted = models.BooleanField(default=False)
-    history = HistoricalRecords(table_name='operational_exception_history')
-    
-    objects = ActiveManager()
+    history = HistoricalRecords(table_name='assignment_rule_resource_group_history')
 
+
+    objects = ActiveManager()
+    
     def __str__(self):
         return str(self.id)
 
     class Meta:
-        db_table = 'operational_exception'
+        db_table = 'assignment_rule_resource_group'
         indexes = [
             models.Index(fields=['id'])
         ]
