+++ backend/api/models/operational_exception_type.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.db import models
 from django.utils import timezone
-from .resourceGroups import ResourceGroups
+from .resource_group import ResourceGroups
 from simple_history.models import HistoricalRecords
 from api.utils.model_manager import ActiveManager
