+++ backend/api/models/resource.py
@@ -1,23 +1,22 @@
 from django.conf import settings
 from django.db import models
 from django.utils import timezone
-from .user import User
-from .scheduleRunStatus import ScheduleRunStatus
+from .resource_group import ResourceGroups
 from simple_history.models import HistoricalRecords
 from api.utils.model_manager import ActiveManager
+from .weekly_shift_template import WeeklyShiftTemplate
 
-
-class ScheduleRun(models.Model):
+class Resources(models.Model):
     id = models.AutoField(primary_key=True)
-    triggered_on = models.DateTimeField(blank=True, null=True)
-    triggered_by = models.ForeignKey(User, on_delete=models.CASCADE, related_name="triggered_by", blank=True, null=True)
-    schedule_status = models.ForeignKey(ScheduleRunStatus, on_delete=models.CASCADE, related_name="schedule_status", blank=True, null=True)
+    name = models.CharField(max_length=100)
+    weekly_shift_template = models.ForeignKey(WeeklyShiftTemplate, on_delete=models.DO_NOTHING, blank=True, null=True)
+    resource_groups_list = models.JSONField(blank=True, null=True)
     
     # Metadata
     created_at = models.DateTimeField(default=timezone.now)
     created_by = models.ForeignKey(
         settings.AUTH_USER_MODEL,
-        related_name="created_schedule_run",
+        related_name="created_resources",
         on_delete=models.SET_NULL,
         null=True,
         blank=True,
@@ -25,7 +24,7 @@ class ScheduleRun(models.Model):
     updated_at = models.DateTimeField(auto_now=True)
     updated_by = models.ForeignKey(
         settings.AUTH_USER_MODEL,
-        related_name="updated_schedule_run",
+        related_name="updated_resources",
         on_delete=models.SET_NULL,
         null=True,
         blank=True,
@@ -33,15 +32,15 @@ class ScheduleRun(models.Model):
     deleted_at = models.DateTimeField(blank=True, null=True)
     is_active = models.BooleanField(default=True)
     is_deleted = models.BooleanField(default=False)
-    history = HistoricalRecords(table_name='schedule_run_history')
-
-    objects = ActiveManager()
+    history = HistoricalRecords(table_name='resource_history')
     
+    objects = ActiveManager()
+
     def __str__(self):
-        return str(self.id)
+        return self.name
 
     class Meta:
-        db_table = 'schedule_run'
+        db_table = 'resource'
         indexes = [
-            models.Index(fields=['id'])
+            models.Index(fields=['id', 'name'])
         ]
