+++ backend/api/models/task_resource_assignment.py
@@ -1,22 +1,22 @@
 from django.conf import settings
 from django.db import models
 from django.utils import timezone
-from .resourceGroups import ResourceGroups
+from .task import Tasks
+from .resource import Resources
 from simple_history.models import HistoricalRecords
 from api.utils.model_manager import ActiveManager
-from .weeklyShiftTemplate import WeeklyShiftTemplate
 
-class Resources(models.Model):
+
+class TasksResourceAssignment(models.Model):
     id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=100)
-    weekly_shift_template = models.ForeignKey(WeeklyShiftTemplate, on_delete=models.DO_NOTHING, blank=True, null=True)
-    resource_groups_list = models.JSONField(blank=True, null=True)
+    task = models.ForeignKey(Tasks, on_delete=models.CASCADE, related_name="task_assignment", blank=True, null=True)
+    resource = models.ForeignKey(Resources, on_delete=models.CASCADE, related_name="resource_assignment", blank=True, null=True)
     
     # Metadata
     created_at = models.DateTimeField(default=timezone.now)
     created_by = models.ForeignKey(
         settings.AUTH_USER_MODEL,
-        related_name="created_resources",
+        related_name="created_task_resource_assignment",
         on_delete=models.SET_NULL,
         null=True,
         blank=True,
@@ -24,7 +24,7 @@ class Resources(models.Model):
     updated_at = models.DateTimeField(auto_now=True)
     updated_by = models.ForeignKey(
         settings.AUTH_USER_MODEL,
-        related_name="updated_resources",
+        related_name="updated_task_resource_assignment",
         on_delete=models.SET_NULL,
         null=True,
         blank=True,
@@ -32,15 +32,15 @@ class Resources(models.Model):
     deleted_at = models.DateTimeField(blank=True, null=True)
     is_active = models.BooleanField(default=True)
     is_deleted = models.BooleanField(default=False)
-    history = HistoricalRecords(table_name='resource_history')
-    
-    objects = ActiveManager()
+    history = HistoricalRecords(table_name='tasks_resource_assignment_history')
 
+    objects = ActiveManager()
+    
     def __str__(self):
-        return self.name
+        return str(self.id)
 
     class Meta:
-        db_table = 'resource'
+        db_table = 'tasks_resource_assignment'
         indexes = [
-            models.Index(fields=['id', 'name'])
+            models.Index(fields=['id'])
         ]
