+++ backend/api/models/weekly_shift_template.py
